move extract and observe to page (#329)
* move act evals to stagehand.page

* add basic act and make act necessary in type

* move extract and observe

* example

* changeset
