+++ .changeset/serious-pets-kiss.md
@@ -1,0 +1,5 @@
+---
+"@browserbasehq/stagehand": patch
+---
+
+moves extract and act -> page and deprecates stagehand.extract and stagehand.observe
