+++ lib/StagehandPage.ts
@@ -7,12 +7,23 @@ import { ActOptions, ActResult, GotoOptions, Stagehand } from "./index";
 import { StagehandActHandler } from "./handlers/actHandler";
 import { StagehandContext } from "./StagehandContext";
 import { Page } from "../types/page";
+import {
+  ExtractOptions,
+  ExtractResult,
+  ObserveOptions,
+  ObserveResult,
+} from "../types/stagehand";
+import { z } from "zod";
+import { StagehandExtractHandler } from "./handlers/extractHandler";
+import { StagehandObserveHandler } from "./handlers/observeHandler";
 
 export class StagehandPage {
   private stagehand: Stagehand;
   private intPage: Page;
   private intContext: StagehandContext;
   private actHandler: StagehandActHandler;
+  private extractHandler: StagehandExtractHandler;
+  private observeHandler: StagehandObserveHandler;
   private llmClient: LLMClient;
 
   constructor(
@@ -36,6 +47,16 @@ export class StagehandPage {
       stagehandPage: this,
       stagehandContext: this.intContext,
       llmClient: llmClient,
+    });
+    this.extractHandler = new StagehandExtractHandler({
+      stagehand: this.stagehand,
+      logger: this.stagehand.logger,
+      stagehandPage: this,
+    });
+    this.observeHandler = new StagehandObserveHandler({
+      stagehand: this.stagehand,
+      logger: this.stagehand.logger,
+      stagehandPage: this,
     });
     this.llmClient = llmClient;
   }
@@ -263,6 +284,154 @@ export class StagehandPage {
           message: `Internal error: Error acting: ${e.message}`,
           action: action,
         };
+      });
+  }
+
+  async extract<T extends z.AnyZodObject>({
+    instruction,
+    schema,
+    modelName,
+    modelClientOptions,
+    domSettleTimeoutMs,
+    useTextExtract,
+  }: ExtractOptions<T>): Promise<ExtractResult<T>> {
+    if (!this.extractHandler) {
+      throw new Error("Extract handler not initialized");
+    }
+
+    const requestId = Math.random().toString(36).substring(2);
+    const llmClient = modelName
+      ? this.stagehand.llmProvider.getClient(modelName, modelClientOptions)
+      : this.llmClient;
+
+    this.stagehand.log({
+      category: "extract",
+      message: "running extract",
+      level: 1,
+      auxiliary: {
+        instruction: {
+          value: instruction,
+          type: "string",
+        },
+        requestId: {
+          value: requestId,
+          type: "string",
+        },
+        modelName: {
+          value: llmClient.modelName,
+          type: "string",
+        },
+      },
+    });
+
+    return this.extractHandler
+      .extract({
+        instruction,
+        schema,
+        llmClient,
+        requestId,
+        domSettleTimeoutMs,
+        useTextExtract,
+      })
+      .catch((e) => {
+        this.stagehand.log({
+          category: "extract",
+          message: "error extracting",
+          level: 1,
+          auxiliary: {
+            error: {
+              value: e.message,
+              type: "string",
+            },
+            trace: {
+              value: e.stack,
+              type: "string",
+            },
+          },
+        });
+
+        if (this.stagehand.enableCaching) {
+          this.stagehand.llmProvider.cleanRequestCache(requestId);
+        }
+
+        throw e;
+      });
+  }
+
+  async observe(options?: ObserveOptions): Promise<ObserveResult[]> {
+    if (!this.observeHandler) {
+      throw new Error("Observe handler not initialized");
+    }
+
+    const requestId = Math.random().toString(36).substring(2);
+    const llmClient = options?.modelName
+      ? this.stagehand.llmProvider.getClient(
+          options.modelName,
+          options.modelClientOptions,
+        )
+      : this.llmClient;
+
+    this.stagehand.log({
+      category: "observe",
+      message: "running observe",
+      level: 1,
+      auxiliary: {
+        instruction: {
+          value: options?.instruction,
+          type: "string",
+        },
+        requestId: {
+          value: requestId,
+          type: "string",
+        },
+        modelName: {
+          value: llmClient.modelName,
+          type: "string",
+        },
+      },
+    });
+
+    return this.observeHandler
+      .observe({
+        instruction:
+          options?.instruction ??
+          "Find actions that can be performed on this page.",
+        llmClient,
+        useVision: options?.useVision ?? false,
+        fullPage: false,
+        requestId,
+        domSettleTimeoutMs: options?.domSettleTimeoutMs,
+      })
+      .catch((e) => {
+        this.stagehand.log({
+          category: "observe",
+          message: "error observing",
+          level: 1,
+          auxiliary: {
+            error: {
+              value: e.message,
+              type: "string",
+            },
+            trace: {
+              value: e.stack,
+              type: "string",
+            },
+            requestId: {
+              value: requestId,
+              type: "string",
+            },
+            instruction: {
+              value: options?.instruction,
+              type: "string",
+            },
+          },
+        });
+
+        if (this.stagehand.enableCaching) {
+          this.stagehand.llmProvider.cleanRequestCache(requestId);
+        }
+
+        throw e;
       });
   }
 }
